add jwt and create api

+++ controllers/api/v1/posts_api.js
@@ -21,16 +21,18 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    //   if (post.user == req.user.id) {
-    post.remove();
-    await Comment.deleteMany({ post: req.params.id });
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: 'Post deleted successfully',
-    });
-    //   } else {
-    //     return res.redirect('back');
-    //   }
+      return res.json(200, {
+        message: 'Post deleted successfully',
+      });
+    } else {
+      return res.status(401).json({
+        message: 'You can not Delete it ',
+      });
+    }
   } catch (err) {
     console.log(err);
     return res.json(500, {
